+++ Conspiratio.Lib/Allgemein/IJaNeinFrage.cs
@@ -1,9 +1,0 @@
-﻿using System.Windows.Forms;
-
-namespace Conspiratio.Lib.Allgemein
-{
-    public interface IJaNeinFrage
-    {
-        DialogResult ShowDialogText(string textFrage, string textJa = "Ja", string textNein = "Nein");
-    }
-}