+++ Conspiratio.Lib/Gameplay/Privilegien/PrivAmtNiederlegen.cs
@@ -1,5 +1,4 @@
-﻿using System.Windows.Forms;
-
+﻿using Conspiratio.Lib.Allgemein;
 using Conspiratio.Lib.Gameplay.Spielwelt;
 
 namespace Conspiratio.Lib.Gameplay.Privilegien
@@ -12,7 +11,7 @@ namespace Conspiratio.Lib.Gameplay.Privilegien
 
         public override void PrivExecute()
         {
-            if (SW.UI.JaNeinFrage.ShowDialogText("Wollt Ihr wirklich\nEuer Amt niederlegen?", "Ja", "Nein") == DialogResult.Yes)
+            if (SW.UI.YesNoQuestion.ShowDialogText("Wollt Ihr wirklich\nEuer Amt niederlegen?", "Ja", "Nein").GetAwaiter().GetResult() == DialogResultGame.Yes)
             {
                 // Absetzungsanträge zurückziehen
                 for (int i = 1; i < SW.Statisch.GetMaxAnzahlAmtsenthebungen(); i++)