+++ Conspiratio.Lib/Gameplay/Spielwelt/StatischeSpieldaten.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 
 using Conspiratio.Lib.Gameplay.Aemter;
 using Conspiratio.Lib.Gameplay.Ereignisse;
@@ -22,6 +23,8 @@ namespace Conspiratio.Lib.Gameplay.Spielwelt
 
         #region Allgemein
 
+        public string SavegamePath { get; } = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "Conspiratio");
+        
         /// <summary>
         /// Gibt die Jahreszahl zu Beginn eines neuen Spiels an.
         /// </summary>
@@ -30,6 +33,7 @@ namespace Conspiratio.Lib.Gameplay.Spielwelt
         /// <summary>
         /// Enthält die Texte aller Tipps.
         /// </summary>
+        // ReSharper disable once CollectionNeverQueried.Global
         public string[] Tipps { get; private set; }
 
         private int TippsMaxIndex = 0;
@@ -227,6 +231,55 @@ namespace Conspiratio.Lib.Gameplay.Spielwelt
         #endregion
 
         #region Initialisieren
+        public void Initialisieren()
+        {
+            var strafarten = new IStrafe[GetMaxAnzahlStrafen()];
+
+            strafarten[0] = new StrafePranger();
+            strafarten[1] = new StrafeKerker();
+            strafarten[2] = new StrafeGeldstrafe();
+            strafarten[3] = new StrafeAmtsenthebung();
+
+            var privilegien = new IPrivileg[GetMaxPriv()];
+
+            privilegien[1] = new PrivMedikus();
+            privilegien[2] = new PrivAmtNiederlegen();
+            privilegien[3] = new PrivTestament();
+            privilegien[4] = new PrivRohstoffrecht();
+            privilegien[5] = new PrivEinkommen();
+            privilegien[6] = new PrivUntergebene();
+            privilegien[7] = new PrivKerkerklatsch();
+            privilegien[8] = new PrivConfessio();
+            privilegien[9] = new PrivProzessInitiieren();
+            privilegien[10] = new PrivBauwerkStiften();
+            privilegien[11] = new PrivHaendler();
+            privilegien[12] = new PrivKaufmann();
+            privilegien[13] = new PrivGroßkaufmann();
+            privilegien[14] = new PrivUmsatzsteuerFestlegen();
+            privilegien[15] = new PrivSparplan();
+            privilegien[16] = new PrivKeinKirchenzehnt();
+            privilegien[17] = new PrivVergifteterWein();
+            privilegien[18] = new PrivWachen();
+            privilegien[19] = new PrivLeibgarde();
+            privilegien[20] = new PrivHenkersHand();
+            privilegien[21] = new PrivKorruptionsgelder();
+            privilegien[22] = new PrivSchmuggel();
+            privilegien[23] = new PrivZollkartell();
+            privilegien[24] = new PrivKirGesetzeAendern();
+            privilegien[25] = new PrivFinGesetzeAendern();
+            privilegien[26] = new PrivStrafGesetzeAendern();
+            privilegien[27] = new PrivSteuerhinterziehungA();
+            privilegien[28] = new PrivSteuerhinterziehungB();
+            privilegien[29] = new PrivSteuerhinterziehungC();
+            privilegien[30] = new PrivGuenstigeKredite();
+            privilegien[31] = new PrivZollfrei();
+            privilegien[32] = new PrivPrediger();
+            privilegien[33] = new PrivFestGeben();
+            privilegien[34] = new PrivJurist();
+
+            Initialisieren(strafarten, privilegien);
+        }
+
         public void Initialisieren(IStrafe[] strafarten, IPrivileg[] privilegien)
         {
             #region Gesetze
@@ -1038,26 +1091,23 @@ namespace Conspiratio.Lib.Gameplay.Spielwelt
             Haeuser = new Haus[MaxHausID];
 
             #region Zustandsbezeichnungen
-            HausZustandsbezeichnung[] aHauszustandBezeichnungenDer;
-            HausZustandsbezeichnung[] aHauszustandBezeichnungenDie;
-            HausZustandsbezeichnung[] aHauszustandBezeichnungenDas;
-            int iMaxZustandsbezeichnungen = 5;
+            int maxZustandsbezeichnungen = 5;
 
-            aHauszustandBezeichnungenDer = new HausZustandsbezeichnung[iMaxZustandsbezeichnungen];
+            var aHauszustandBezeichnungenDer = new HausZustandsbezeichnung[maxZustandsbezeichnungen];
             aHauszustandBezeichnungenDer[0] = new HausZustandsbezeichnung("prächtiger", 80, 100);
             aHauszustandBezeichnungenDer[1] = new HausZustandsbezeichnung("schöner", 60, 79);
             aHauszustandBezeichnungenDer[2] = new HausZustandsbezeichnung("gut erhaltener", 40, 59);
             aHauszustandBezeichnungenDer[3] = new HausZustandsbezeichnung("verwahrloster", 20, 39);
             aHauszustandBezeichnungenDer[4] = new HausZustandsbezeichnung("verfallener", 0, 19);
 
-            aHauszustandBezeichnungenDie = new HausZustandsbezeichnung[iMaxZustandsbezeichnungen];
+            var aHauszustandBezeichnungenDie = new HausZustandsbezeichnung[maxZustandsbezeichnungen];
             aHauszustandBezeichnungenDie[0] = new HausZustandsbezeichnung("prächtige", 80, 100);
             aHauszustandBezeichnungenDie[1] = new HausZustandsbezeichnung("schöne", 60, 79);
             aHauszustandBezeichnungenDie[2] = new HausZustandsbezeichnung("gut erhaltene", 40, 59);
             aHauszustandBezeichnungenDie[3] = new HausZustandsbezeichnung("verwahrloste", 20, 39);
             aHauszustandBezeichnungenDie[4] = new HausZustandsbezeichnung("verfallene", 0, 19);
 
-            aHauszustandBezeichnungenDas = new HausZustandsbezeichnung[iMaxZustandsbezeichnungen];
+            var aHauszustandBezeichnungenDas = new HausZustandsbezeichnung[maxZustandsbezeichnungen];
             aHauszustandBezeichnungenDas[0] = new HausZustandsbezeichnung("prächtiges", 80, 100);
             aHauszustandBezeichnungenDas[1] = new HausZustandsbezeichnung("schönes", 60, 79);
             aHauszustandBezeichnungenDas[2] = new HausZustandsbezeichnung("gut erhaltenes", 40, 59);
@@ -1138,7 +1188,9 @@ namespace Conspiratio.Lib.Gameplay.Spielwelt
 
             #region Privilegien anlegen
 
-            Privilegien = privilegien;  // Von außen als Interface übergeben, da die eigentliche Implementierung der einzelnen Privilegien im jeweiligen Client und nicht in der Lib existiert
+            // Ursprünglich: Von außen als Interface übergeben, da die eigentliche Implementierung der einzelnen Privilegien im jeweiligen Client und nicht in der Lib existierte
+            // Aktuell: Implementierung existiert hier in der Lib, aber das Prinzip so gelassen aus Gründen der Testbarkeit und Verminderung fester Abhängigkeiten (dependency injection) 
+            Privilegien = privilegien;  
 
             #endregion
 
@@ -1209,7 +1261,9 @@ namespace Conspiratio.Lib.Gameplay.Spielwelt
 
             #region Strafarten
 
-            Strafarten = strafarten;  // Von außen als Interface übergeben, da die eigentliche Implementierung der einzelnen Strafen im jeweiligen Client und nicht in der Lib existiert
+            // Ursprünglich: Von außen als Interface übergeben, da die eigentliche Implementierung der einzelnen Privilegien im jeweiligen Client und nicht in der Lib existierte
+            // Aktuell: Implementierung existiert hier in der Lib, aber das Prinzip so gelassen aus Gründen der Testbarkeit und Verminderung fester Abhängigkeiten (dependency injection) 
+            Strafarten = strafarten;
 
             #endregion